Deleted ts airtable files except interface. Moved to lib folder
