+++ src/lib/utils/customerUtils.js
