+++ src/screens/Customers/CustomerMain.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
-import { CustomerRecord } from '../../utils/airtable/interface';
-import { getAllCustomers } from '../../utils/airtable/requests';
+import { CustomerRecord } from '../../lib/airtable/interface';
+import { getAllCustomers } from '../../lib/airtable/request';
 import { Link, RouteComponentProps } from 'react-router-dom';
 
 function CustomerMain(props: RouteComponentProps) {
